+++ ui/utils/stringToUtf8Array.mjs
