+++ ui/verify.mjs
@@ -1,9 +1,9 @@
 import { array_to_hex } from "../lib-mlsag-js/bytes.mjs";
 import { LSAG_Verify, keccak } from "../lib-mlsag-wasm/index.mjs";
 import { byId } from "./byId.mjs";
-import { getMessageHash } from "./getMessageHash.mjs";
+import { getMessageHash } from "./utils/getMessageHash.mjs";
 import { ring_pub_keys_placeholder_text } from "./signverify.mjs";
-import { parsePublicKeys } from "./parsePublicKeys.mjs";
+import { parsePublicKeys } from "./utils/parsePublicKeys.mjs";
 
 byId("ring_pubkeys").setAttribute(
   "placeholder",
